DRR - Cpptraj: Fix potential exception when reading large (>9 digit) integers from arg list (thanks to JMS for report). Also add 'last' and 'lastframe' keywords to trajin command for selecting last frame with offset and just last frame respectively.

+++ src/ArgList.cpp
@@ -302,10 +302,10 @@ public:
 /// Convert the input string to an integer.
 inline int convertToInteger(string const &s) {
   istringstream iss(s);
-  int i;
+  long int i;
   if (!(iss >> i))
     throw BadConversion("convertToInteger(\"" + s + "\")");
-  return i;
+  return (int)i;
 }
 
 // convertToDouble()
